MT_000 Добавлена библиотека для логирования, поправил структуру RoleService
