+++ src/main/java/com/springboot/service/UserService.java
@@ -14,8 +14,11 @@ public interface UserService {
     void saveUser(User user);
 
     List<User> getAllUsers();
+
     User findUserByFirstname(String firstname);
+
     User getUserByEmail(String email);
+
     User getUserById(Long id);
 
     void update(User user);
@@ -23,7 +26,9 @@ public interface UserService {
     void deleteById(Long id);
 
     Role getRole(String rolename);
+
     void setRole(User user, String role);
+
     void setDefaultRole(User user);
 
 }
